다국어 제목 지원, 버그 수정

+++ baekjoon/judge.py
@@ -109,6 +109,9 @@ def generate_code(args):
     ext = path.splitext(template_path)[-1]
     assert ext != "", "[!] File extension not found"
     new_path = prob_dir + sep + str(pid) + ext
+    if os.path.exists(new_path):
+        print("[!] File exists:", new_path)
+        return
     inf = open(template_path, 'r')
     outf = open(new_path, 'w')
     for line in inf:
